Split up ImmutableSortedDictionary types
